Backend complete version 1

+++ springapp/src/main/java/com/examly/springapp/services/EmployeeService.java
@@ -2,6 +2,7 @@ package com.examly.springapp.services;
 
 import com.examly.springapp.model.Employee;
 import com.examly.springapp.repository.EmployeeRepository;
+import com.examly.springapp.repository.RouteRepository;
 import java.util.ArrayList;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,6 +11,7 @@ import org.springframework.stereotype.Service;
 @Service
 public class EmployeeService {
   @Autowired private EmployeeRepository employeeRepository;
+  @Autowired private RouteRepository routeRepository;
 
   public List<Employee> getEmployee() {
     Iterable<Employee> employees = employeeRepository.findAll();
@@ -42,6 +44,10 @@ public class EmployeeService {
   public void deleteEmployee(int id) {
     if (!employeeRepository.existsById(id))
       throw new IllegalArgumentException("User Already Exits");
+
+    if (routeRepository.existsByEmployeeId(id))
+      throw new IllegalArgumentException(
+          "Remove the Employee from route first, then delete the account");
     employeeRepository.deleteById(id);
   }
 }
